chat-push-message/pyproject.toml
[project]
name = "practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ruff>=0.12.8",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "gradio==5.41.1",
    "requests>=2.32.4",
    "openai-agents>=0.2.5",
    "rich>=13.9.4",
]
